fix | added selected item null checks

+++ source/DevChecker/Tabs/ConfigMgr/CmApplicatonsTableViewer.cs
@@ -49,7 +49,7 @@ namespace DevChecker.Tabs.ConfigMgr
         private async void onInstallClicked(object sender, RoutedEventArgs e)
         {
             var selected = (ConfigMgrHelpers.Deploy.Application)this.DataGrid.SelectedItem;
-            if (MessageBox.Show("Are you sure you want to install " + selected.Name + "?", "Install application", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+            if (selected != null && MessageBox.Show("Are you sure you want to install " + selected.Name + "?", "Install application", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
             {
                 await selected.InstallAsync();
             }
@@ -58,7 +58,7 @@ namespace DevChecker.Tabs.ConfigMgr
         private async void onUninstallClicked(object sender, RoutedEventArgs e)
         {
             var selected = (ConfigMgrHelpers.Deploy.Application)this.DataGrid.SelectedItem;
-            if (MessageBox.Show("Are you sure you want to uninstall " + selected.Name + "?", "Uninstall application", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+            if (selected != null && MessageBox.Show("Are you sure you want to uninstall " + selected.Name + "?", "Uninstall application", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
             {
                 await selected.UninstallAsync();
             }
